set exit code in fixer script

+++ scripts/fix-non-ascii-quotes.py
@@ -10,9 +10,15 @@ filename = sys.argv[1]
 
 with open(filename, "r+") as f:
     contents = f.read()
-    fixed_contents = re.sub(unicode_single_quotes_pattern, "'", contents)
-    fixed_contents = re.sub(unicode_double_quotes_pattern, '\\"', fixed_contents)
+    fixed_contents = unicode_single_quotes_pattern.sub("'", contents)
+    fixed_contents = unicode_double_quotes_pattern.sub('\\"', fixed_contents)
+    if fixed_contents == contents:
+        # nothing to fix we're done
+        sys.exit(0)
     # overwrite the data in the file
     f.seek(0)
     f.write(fixed_contents)
     f.truncate()
+print(f"fixed non-ascii quotes in {filename}")
+# need to exit 1 to tell pre-commit that we fixed something
+sys.exit(1)
